ml_service/requirements.txt
flask==3.1.0
flask-cors==5.0.1
numpy==1.26.4
Pillow==11.1.0
tensorflow==2.19.0
